// client/components/atoms/ToastModel.atom.tsx
import styled from "styled-components";

export const ToastModal = styled.div`
  display: flex;
  flex-direction: column;
  align-items: center;
  text-align: center;
  background: #fff;
  font-size: 14px;
  color: #333;
  padding: 14px;
  border-radius: 8px;
  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
  max-width: 400px;
  margin: 0 auto;
`;