// client/components/atoms/Button.atom.tsx
import styled from "styled-components";

export const Button = styled.button`
  padding: 0.8rem 0.9rem 0.8rem 0.9rem;
`;

export const BlueButton = styled(Button)`
  background-color: #4285f4;
  color: white;
  border: none;
  font-size: 14px;
  font-weight: bold;
  border-radius: 4px;
  cursor: pointer;
`;